# docs/protocol/openapi.yaml
openapi: 3.0.3
info:
  title: Agent-Native IM API
  version: 0.1.0
  description: |
    Agent-Native IM platform API. Designed for AI agent self-integration.

    ## Authentication
    - **Web users**: `POST /api/v1/auth/login` → JWT token
    - **Bots**: Use the token returned when creating a bot via dashboard

    Both use `Authorization: Bearer <token>` header.

    ## Message Layers
    Messages have multiple layers for different audiences:
    - `thinking`: reasoning chain (collapsible in UI)
    - `status`: progress updates `{phase, progress, text}`
    - `data`: structured JSON for agent consumption
    - `summary`: natural language for human reading
    - `interaction`: interactive cards (approval/choice/form)
servers:
  - url: http://localhost:9800
    description: Local development
components:
  securitySchemes:
    userAuth:
      type: http
      scheme: bearer
      description: JWT token from /auth/login
    botAuth:
      type: http
      scheme: bearer
      description: Bot API token from bot creation
  schemas:
    Response:
      type: object
      properties:
        ok:
          type: boolean
        data:
          description: Response payload (varies by endpoint)
        error:
          type: string
          description: Error message (only when ok=false)
...
    ReleaseSection:
      type: object
      properties:
        kind:
          type: string
          enum: [feature, fix, breaking, security, developer, known_issue]
        title:
          type: string
        items:
          type: array
          items:
            type: string
    ReleaseAction:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        command:
          type: string
        url:
          type: string
        required:
          type: boolean
    Release:
      type: object
      properties:
        id:
          type: integer
          format: int64
        public_id:
          type: string
          format: uuid
        version:
          type: string
        title:
          type: string
        summary:
          type: string
        component:
          type: string
          enum: [platform, web, mobile, openclaw, zebra, hermes, sdk, docs]
        platform:
          type: string
          enum: [all, web, ios, android, desktop, agent]
        channel:
          type: string
          enum: [production, beta, internal]
        sections:
          type: array
          items:
            $ref: '#/components/schemas/ReleaseSection'
        required_actions:
          type: array
          items:
            $ref: '#/components/schemas/ReleaseAction'
        known_issues:
          type: array
          items:
            type: string
        published_at:
          type: string
          format: date-time
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        is_read:
          type: boolean
        unread_count:
          type: integer
    ReleaseListResponse:
      type: object
      properties:
        releases:
          type: array
          items:
            $ref: '#/components/schemas/Release'
        unread_count:
          type: integer
paths:
...
  /api/v1/releases:
    get:
      summary: List release notes
      operationId: listReleases
      security:
        - userAuth: []
        - botAuth: []
      parameters:
        - name: channel
          in: query
          schema:
            type: string
            default: production
        - name: component
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        '200':
          description: Release notes with current user's unread count.
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Response'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/ReleaseListResponse'
  /api/v1/releases/latest:
    get:
      summary: Get latest release note
      operationId: getLatestRelease
      security:
        - userAuth: []
        - botAuth: []
      parameters:
        - name: channel
          in: query
          schema:
            type: string
            default: production
      responses:
        '200':
          description: Latest release note for the selected channel.
          content:
            application/json:
              example:
                ok: true
                data:
                  release:
                    id: 1
                    version: "2026.5.14"
                  unread_count: 1
  /api/v1/releases/{id}/read:
    post:
      summary: Mark a release note as read
      operationId: markReleaseRead
      security:
        - userAuth: []
        - botAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Release read marker upserted.
...